+++ src/aaagent/__init__.py
@@ -5,6 +5,7 @@ from aaagent.core.message import Message
 from aaagent.core.plugin import PluginContext
 from aaagent.core.session import Session, SessionStore
 from aaagent.core.types import LLMProvider, PROVIDER_TYPE_REGISTRY, register_provider_type
+from aaagent.testing import FakeProvider
 
 __all__ = [
     "AgentContext",
@@ -12,6 +13,7 @@ __all__ = [
     "Application",
     "DefaultAgentLoop",
     "EventBus",
+    "FakeProvider",
     "Message",
     "PluginContext",
     "Session",